Move update_artists() to utils.py
