+++ drawArt.py
@@ -1,6 +1,5 @@
 import sys
 import utils
-import numpy as np
 from matplotlib import pyplot as plt
 from matplotlib import animation
 from functools import partial
@@ -47,46 +46,9 @@ artists = lines + circles + [outline]
 
 
 ## animation
-def update_artists(frame_num):
-    current_phases_radian = float(frame_num)/num_frames_per_cycle*2*np.pi*rotation_speeds + start_phases
-    x_centers_circle, y_centers_circle = get_circle_centers(circle_radii, current_phases_radian)
-    update_lines(lines, x_centers_circle, y_centers_circle)
-    update_circles(circles, circle_radii, x_centers_circle, y_centers_circle)
-    update_outline(outline, x_centers_circle[-1], y_centers_circle[-1])
-    return artists
-
-
-def get_circle_centers(radii, phases_radian):
-    x_line_endpoints = np.cos(phases_radian) * radii
-    y_line_endpoints = np.sin(phases_radian) * radii
-    # add the lines end to end to find the circle centers
-    x_centers_circle = np.concatenate(([0], np.cumsum(x_line_endpoints)))
-    y_centers_circle = np.concatenate(([0], np.cumsum(y_line_endpoints)))
-    return x_centers_circle, y_centers_circle
-
-
-def update_lines(lines, x_line_endpoints, y_line_endpoints):
-    for idx, line in enumerate(lines):
-        x_line = x_line_endpoints[idx:idx+2]
-        y_line = y_line_endpoints[idx:idx+2]
-        line.set_data(x_line, y_line)
-
-
-def update_circles(circles, circle_radii, x_centers_circle, y_centers_circle):
-    for idx, circle in enumerate(circles):
-        x_coords, y_coords = utils.get_circle_coordinates(x_centers_circle[idx], y_centers_circle[idx], circle_radii[idx])
-        circle.set_data(x_coords, y_coords)
-
-
-def update_outline(outline, x_pos, y_pos):
-    x_outline = np.concatenate((outline.get_xdata(), np.array(x_pos, ndmin=1)))
-    y_outline = np.concatenate((outline.get_ydata(), np.array(y_pos, ndmin=1)))
-    outline.set_data(x_outline, y_outline)
-
-
 anim = animation.FuncAnimation(
     fig,
-    update_artists,
+    partial(utils.update_artists, lines, circles, outline, rotation_speeds, start_phases, circle_radii, num_frames_per_cycle),
     init_func=partial(utils.initialize_artists, artists),
     frames=num_frames,
     interval=20,
